inventory-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
aiopg==1.4.0
redis[asyncio]==5.0.1
aiokafka==0.10.0
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
setuptools<78

opentelemetry-distro==0.44b0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-sqlalchemy==0.44b0
opentelemetry-instrumentation-redis==0.44b0
opentelemetry-instrumentation-httpx==0.44b0
opentelemetry-instrumentation-logging==0.44b0

prometheus-client==0.20.0
